Implement better id encoding.

+++ src/rust/ensogl/src/display/world.rs
@@ -142,7 +142,7 @@ impl World {
 
     fn init_composer(&self) {
         let mouse_hover_ids     = self.scene.mouse.hover_ids.clone_ref();
-        let mut pixel_read_pass = PixelReadPass::<f32>::new(&self.scene.mouse.position);
+        let mut pixel_read_pass = PixelReadPass::<u8>::new(&self.scene.mouse.position);
         pixel_read_pass.set_callback(move |v| {
             mouse_hover_ids.set(Vector4::from_iterator(v.iter().map(|value| *value as u32)))
         });
